Clean up warnings in graphics pipelines
Cleans up loss of precision warnings in
graphics pipelines.

+++ GraphicsPipelinePBR.cpp
@@ -33,7 +33,7 @@ void GraphicsPipelinePBR::createDescriptorPool(LogicalDevice& logicalDevice)
 
     poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
     poolInfo.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
-    poolInfo.poolSizeCount = poolSizes.size();
+    poolInfo.poolSizeCount = static_cast<uint32_t>(poolSizes.size());
     poolInfo.pPoolSizes = poolSizes.data();
     poolInfo.maxSets = 3;
 
@@ -116,7 +116,7 @@ void GraphicsPipelinePBR::createDescriptorSetLayout(LogicalDevice& logicalDevice
     bindings[10].stageFlags = VK_SHADER_STAGE_FRAGMENT_BIT;
     
     layoutInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-    layoutInfo.bindingCount = bindings.size();
+    layoutInfo.bindingCount = static_cast<uint32_t>(bindings.size());
     layoutInfo.pBindings = bindings.data();
 
     if (VK_SUCCESS != vkCreateDescriptorSetLayout(logicalDevice.getDevice(), &layoutInfo, nullptr, &descriptorSetLayout))
